% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_log.R
\name{read_log}
\alias{read_log}
\title{read_log}
\usage{
read_log(x)
}
\arguments{
\item{x}{A string with the path to the Telegram group log file in JSON format}
}
\description{
This function read the Telegram group log file (only in JSON format) and create three dataframes: 1. full data, 2. messages data, 3. services data.
}
