% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_messages.R, R/users_messages.R
\name{users_messages}
\alias{users_messages}
\title{daily_messages}
\usage{
users_messages(messages_dataframe)

users_messages(messages_dataframe)
}
\arguments{
\item{messages_dataframe}{A messages data.frame obtained from the Telegram group file}

\item{messages_dataframe}{A messages data.frame obtained from the Telegram group file}
}
\description{
This function obtain a data.frame with the number of daily messages.

This function obtain a data.frame with the users and its number of messages.
}
