% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_messages.R
\name{users_messages}
\alias{users_messages}
\title{users_messages}
\usage{
users_messages(messages_dataframe)
}
\arguments{
\item{messages_dataframe}{A messages data.frame obtained from the Telegram group fil}
}
\description{
This function obtaint a data.frame with the users and its number of messages.
}
